Huge refactor part 2
