--- ift6758/data/question_1.py
@@ -1,40 +1,0 @@
-import pandas as pd
-
-def get_player_stats(year: int, player_type: str) -> pd.DataFrame:
-    """
-
-    Uses Pandas' built in HTML parser to scrape the tabular player statistics from
-    https://www.hockey-reference.com/leagues/ . If the player played on multiple 
-    teams in a single season, the individual team's statistics are discarded and
-    the total ('TOT') statistics are retained (the multiple team names are discarded)
-
-    Args:
-        year (int): The first year of the season to retrieve, i.e. for the 2016-17
-            season you'd put in 2016
-        player_type (str): Either 'skaters' for forwards and defensemen, or 'goalies'
-            for goaltenders.
-    """
-
-    if player_type not in ["skaters", "goalies"]:
-        raise RuntimeError("'player_type' must be either 'skaters' or 'goalies'")
-    
-    url = f'https://www.hockey-reference.com/leagues/NHL_{year}_{player_type}.html'
-
-    print(f"Retrieving data from '{url}'...")
-
-    # Use Pandas' built in HTML parser to retrieve the tabular data from the web data
-    # Uses BeautifulSoup4 in the background to do the heavylifting
-    df = pd.read_html(url, header=1)[0]
-
-    # get players which changed teams during a season
-    players_multiple_teams = df[df['Tm'].isin(['TOT'])]
-
-    # filter out players who played on multiple teams
-    df = df[~df['Player'].isin(players_multiple_teams['Player'])]
-    df = df[df['Player'] != "Player"]
-
-    # add the aggregate rows
-    df = pd.concat([df, players_multiple_teams], ignore_index=True)
-    print(df)
-
-    return df
