+++ setup.py
@@ -1,7 +1,7 @@
 from setuptools import find_packages, setup
 
 setup(
-    name='ift6758',
+    name='src',
     packages=find_packages(),
     # package_dir={'': 'src'},
     version='0.1.0',
